e2e jaeger (cloudflare still broken)
