+++ internal/e2e/src/index.ts
@@ -1,2 +1,3 @@
 export * from './tenv';
 export * from './timeout';
+export * from './jaeger';
